requirements.txt
selenium
discord.py
clipboard